+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `IPublishAutoCAD`
 - Add `IAutoCADPackageBuilder` and `IHazAutoCADPackageBuilder`.
 - Add `AutoCADExtension` to detect AutoCAD version.
-- Add `AutoCADContentsBuilder` to build AutoCAD `AppBundle`.
+- Add `AutoCADContentsBuilder` to build AutoCAD `AppBundle`
+- Add `IssAppBundleBuilder` and `IssAutoCADBuilder`.
+- Update `IssRevitBuilder` to inherit from `IssPackageBuilder`.
 
 ## [1.8.3] / 2025-03-20
 ### Updated